Fix: logger names
