+++ src/main/modules/statementOfMeans/disabilityService.ts
@@ -4,7 +4,7 @@ import {StatementOfMeans} from '../../common/models/statementOfMeans';
 import {Claim} from '../../common/models/claim';
 
 const {Logger} = require('@hmcts/nodejs-logging');
-const logger = Logger.getLogger('dependantsService');
+const logger = Logger.getLogger('disabilityService');
 const disability = new Disability();
 
 export class DisabilityService {
