+++ src/main/modules/statementOfMeans/severeDisabilityService.ts
@@ -4,7 +4,7 @@ import {StatementOfMeans} from '../../common/models/statementOfMeans';
 import {Claim} from '../../common/models/claim';
 
 const {Logger} = require('@hmcts/nodejs-logging');
-const logger = Logger.getLogger('dependantsService');
+const logger = Logger.getLogger('severeDisabilityService');
 const severeDisability = new SevereDisability();
 
 export class SevereDisabilityService {
