+++ src/main/routes/features/response/statementOfMeans/disabilityController.ts
@@ -1,14 +1,14 @@
 import * as express from 'express';
-import { CITIZEN_DISABILITY_URL, CITIZEN_RESIDENCE_URL, CITIZEN_SEVERELY_DISABLED_URL } from '../../../urls';
-import { Disability } from '../../../../common/form/models/statementOfMeans/disability';
-import { ValidationError, Validator } from 'class-validator';
-import { DisabilityService } from '../../../../modules/statementOfMeans/disabilityService';
-import { constructResponseUrlWithIdParams } from '../../../../common/utils/urlFormatter';
+import {CITIZEN_DISABILITY_URL, CITIZEN_RESIDENCE_URL, CITIZEN_SEVERELY_DISABLED_URL} from '../../../urls';
+import {Disability} from '../../../../common/form/models/statementOfMeans/disability';
+import {ValidationError, Validator} from 'class-validator';
+import {DisabilityService} from '../../../../modules/statementOfMeans/disabilityService';
+import {constructResponseUrlWithIdParams} from '../../../../common/utils/urlFormatter';
 
 const citizenDisabilityViewPath = 'features/response/statementOfMeans/disability';
 const router = express.Router();
-const { Logger } = require('@hmcts/nodejs-logging');
-const logger = Logger.getLogger('disabilityService');
+const {Logger} = require('@hmcts/nodejs-logging');
+const logger = Logger.getLogger('disabilityController');
 const disabilityService = new DisabilityService();
 const validator = new Validator();
