+++ src/main/routes/features/response/statementOfMeans/partner/partnerSevereDisabilityController.ts
@@ -1,19 +1,15 @@
 import * as express from 'express';
-import { CITIZEN_DEPENDANTS_URL, CITIZEN_PARTNER_SEVERE_DISABILITY_URL } from '../../../../urls';
-import {
-  PartnerSevereDisability,
-} from '../../../../../common/form/models/statementOfMeans/partner/partnerSevereDisability';
-import { ValidationError, Validator } from 'class-validator';
-import {
-  PartnerSevereDisabilityService,
-} from '../../../../../modules/statementOfMeans/partner/partnerSevereDisabilityService';
-import { constructResponseUrlWithIdParams } from '../../../../../common/utils/urlFormatter';
+import {CITIZEN_DEPENDANTS_URL, CITIZEN_PARTNER_SEVERE_DISABILITY_URL} from '../../../../urls';
+import {PartnerSevereDisability} from '../../../../../common/form/models/statementOfMeans/partner/partnerSevereDisability';
+import {ValidationError, Validator} from 'class-validator';
+import {PartnerSevereDisabilityService} from '../../../../../modules/statementOfMeans/partner/partnerSevereDisabilityService';
+import {constructResponseUrlWithIdParams} from '../../../../../common/utils/urlFormatter';
 
 const partnerViewPath = 'features/response/statementOfMeans/partner/partner-severe-disability';
 const router = express.Router();
 const partnerSevereDisabilityService = new PartnerSevereDisabilityService();
-const { Logger } = require('@hmcts/nodejs-logging');
-const logger = Logger.getLogger('partnerSevereDisabilityService');
+const {Logger} = require('@hmcts/nodejs-logging');
+const logger = Logger.getLogger('partnerSevereDisabilityController');
 const validator = new Validator();
 
 function renderView(form: PartnerSevereDisability, res: express.Response): void {
